Spread the approximated boundaries, to make sure they are all distinct. Fall back to a uniform distribution in case the spreading of the boundaries doesn't fit within the oracle depth limit and the upper bound of 1/1 for all probabilities. Moved some code from Optimizer#consumeEvent to Optimizer#optimize. Added a call to ItemCountsOptimizer#optimize in Optimizer#optimize. Added command-line argument parsing to "file-example-build-and-run.sh". Added command-line option --limit (or -l) to "file-example-build-and-run.sh". Refactored and extended TestItemCountsOptimizer to also test the case where the optimizer has to fall back to producing a uniform distribution.

+++ src/main/scala/org/leialearns/crystallize/util/Oracle.scala
@@ -15,7 +15,7 @@ object Oracle extends Logging {
     value.r.compareTo(item)
   }
 
-  def withinBounds(limit: Int)(value: OrderedRational): Boolean = value.limit < limit
+  def withinBounds(limit: Long)(value: OrderedRational): Boolean = value.limit < limit
 
   def getPrecomputed(resourcePath: String): Map[Long, OrderedRational] = {
     val specRe = "^[(]([0-9]*),([0-9]*) % ([0-9]*)[)]$".r
